+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Prevent the editor from serializing too much information with the filters
 - Add advanced options for the numeric filters
 - Add advanced options for the temporal filters
+- Add advanced options for the textual filters
 
 ## v1.3.8 - 17/12/2018
 - Fix a bug where the external title and caption would be overriden when the component would fetch the widget